refactor: User 검증 단계 추가

+++ src/main/java/com/sangwool/boardproject/usecase/DeleteServiceImpl.java
@@ -1,5 +1,8 @@
 package com.sangwool.boardproject.usecase;
 
+import com.sangwool.boardproject.dto.BoardDeleteDto;
+import com.sangwool.boardproject.dto.CommentDeleteDto;
+import com.sangwool.boardproject.dto.NestedCommentDeleteDto;
 import com.sangwool.boardproject.entity.Comment;
 import com.sangwool.boardproject.entity.NestedComment;
 import com.sangwool.boardproject.repository.BoardRepository;
@@ -20,11 +23,18 @@ public class DeleteServiceImpl implements DeleteService {
     private final NestedCommentRepository nestedCommentRepository;
 
     @Override
-    public Boolean boardDelete(Long boardSeq) {
+    public Boolean boardDelete(BoardDeleteDto boardDeleteDto) {
 
-        // 해당 게시글 번호로 된 모든 댓글을 삭제한다.
-        List<Comment> comments = commentRepository.findAllByBoardSeqNum(boardSeq);
-        comments.forEach(comment -> commentDelete(comment.getCommentSeq()));
+        // 해당 게시글 유효성 검증
+        Long boardSeq = boardDeleteDto.getBoardSeq();
+        if (!boardRepository.findByBoardSeqNum(boardSeq).getUserSeq().equals(boardDeleteDto.getUserSeq())) {
+            return false;
+        }
+
+        // 해당 게시글의 대댓글 삭제
+        List<Comment> comments = commentRepository.findAllByBoardSeqNum(boardDeleteDto.getBoardSeq());
+        comments.forEach(comment ->
+                commentDelete(CommentDeleteDto.buildDto(comment.getCommentSeq(), comment.getUserSeq())));
 
         // 게시글을 삭제한다.
         boardRepository.deleteBoard(boardSeq);
@@ -32,11 +42,18 @@ public class DeleteServiceImpl implements DeleteService {
     }
 
     @Override
-    public Boolean commentDelete(Long commentSeq) {
+    public Boolean commentDelete(CommentDeleteDto commentDeleteDto) {
+
+        // 해당 댓글 권한 검증
+        Long commentSeq = commentDeleteDto.getCommentSeq();
+        if (!commentRepository.findByCommentSeqNum(commentSeq).getUserSeq().equals(commentDeleteDto.getUserSeq())) {
+            return false;
+        }
 
         // 해당 댓글의 대댓글 삭제
         List<NestedComment> nestedComments = nestedCommentRepository.findAllByCommentSeqNum(commentSeq);
-        nestedComments.forEach(nestedComment -> nestedCommentDelete(nestedComment.getNestedCommentSeq()));
+        nestedComments.forEach(nestedComment ->
+                nestedCommentDelete(NestedCommentDeleteDto.buildDto(nestedComment.getNestedCommentSeq(), nestedComment.getUserSeq())));
 
         // 해당 댓글 삭제
         commentRepository.deleteComment(commentSeq);
@@ -44,8 +61,15 @@ public class DeleteServiceImpl implements DeleteService {
     }
 
     @Override
-    public void nestedCommentDelete(Long nestedCommentSeq) {
+    public void nestedCommentDelete(NestedCommentDeleteDto nestedCommentDeleteDto) {
 
+        // 해당 대댓글 권한 검증
+        Long nestedCommentSeq = nestedCommentDeleteDto.getNestedCommentSeq();
+        if (!nestedCommentRepository.findByCommentSeqNum(nestedCommentSeq).getUserSeq().equals(nestedCommentDeleteDto.getUserSeq())) {
+            return;
+        }
+
+        // 해당 대댓글 삭제
         nestedCommentRepository.deleteNestedComment(nestedCommentSeq);
     }
 }
